Moved Spotify OAuth logic to oauth2 module
